refactor: modify get refresh token api by redis
